Temporary revert of r101474, will re-land tomorrow with a change to disable use of precompiled headers in Release builds as our Canary builds are failing with out of memory messages.
Original change description: Switch on use of precompiled headers for VS 2008.

The precompiled header brings in more definitions of abs() for the base/time_win.cc file, and a hint to the compiler was needed to let VS 2008 build the file correctly.

On VS 2008 only, the mini_installer target did not link correctly with precompiled headers enabled so this change disables them for that project.

BUG=none
TEST=it builds (esp. on Windows), existing tests pass

Review URL: http://codereview.chromium.org/7910002

TBR=laforge@chromium.org
Review URL: http://codereview.chromium.org/7945017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@101836 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/time_win.cc
@@ -363,12 +363,7 @@ class HighResNowSingleton {
     if (!IsUsingHighResClock())
       return 0;
 
-    // The static_cast<long> is needed as a hint to VS 2008 to tell it
-    // which version of abs() to use. Other compilers don't seem to
-    // need it, including VS 2010, but to keep code identical we use it
-    // everywhere.
-    // TODO(joi): Remove the hint if/when we no longer support VS 2008.
-    return abs(static_cast<long>((UnreliableNow() - ReliableNow()) - skew_));
+    return abs((UnreliableNow() - ReliableNow()) - skew_);
   }
 
  private:
